+++ src/gnatss/configs/solver.py
