+++ src/gnatss/harmonic_mean.py
